pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "vxtrace"
version = "0.1.0"
description = "Metadata investigation toolkit for researchers and analysts"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
  {name = "VxCypher"}
]
dependencies = [
  "Pillow>=10.0.0",
  "PyPDF2>=3.0.0",
  "python-docx>=1.1.0",
  "openpyxl>=3.1.0",
  "click>=8.1.0",
]

[project.scripts]
vxtrace = "vxtrace.cli:main"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
include = ["vxtrace*"]
